Last fixes

+++ src/aux/free_all.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   free_all.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: clumertz <clumertz@student.42porto.com>    +#+  +:+       +#+        */
+/*   By: dximenes <dximenes@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/20 11:22:04 by dximenes          #+#    #+#             */
-/*   Updated: 2025/11/20 19:14:54 by clumertz         ###   ########.fr       */
+/*   Updated: 2025/11/22 14:27:33 by dximenes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,10 @@ void	free_db_str(char **db_str)
 void	free_node(t_btree *node)
 {
 	if (node->files.in.operator == DOUBLE_ARROW_LEFT)
-		unlink(node->files.in.name);
+	{
+		if (access(node->files.in.name, F_OK) > 0)
+			unlink(node->files.in.name);
+	}
 	if (node->files.in.exists)
 		free(node->files.in.name);
 	if (node->files.out.exists)
